Fix sites/test_ec2 failure when aws user opts specified
Previously, if no AWS user opts were specified, then
the tests in sites/test_ec2.py were skipped (due to a
key error on importing the ec2 config); but if AWS
user opts *were* specified, then the tests would run
in non-local mode.

This commit changes the test so that it runs with a
locally defined config.

I have not verified if the test runs in local mode with
AWS user opts specified: in general this is very awkward
to configure.

This looks like it was messed up in a combination of
PR #1210 (which replaced the local mark with a noci mark)
and PR #1230 (which remoed the noci mark)